Report from the setbuilder's build config.
Refactor the reporter to allow the setbuilder to use its build config
rather than regenerating the configuration from the configuration file.
Using the config file and the build macros exposed an issue if a
macro was undefined that was defined in a build set above the
config file. Using the build set's configuration as used to build
is a better solution.

The reporter was refactored to allow a config class to be used
to report.

The setbuild can now take a configuration file as an input file.