Add an RTEMS config wrapper around Makefile.inc.
This is an internal command that wraps the Makefile.inc file in
a BSP so you can get at the configuration values from the
command line without knowing anything about Makefile.inc.

To use:

 $ ../source-builder/sb-rtems-config \
     --rtems=$HOME/development/rtems/4.11 \
     --rtems-bsp=sis \
      rtems_cpu_cflags

The option '--list' lists all available configuration values.
The configuration values are match case insensitive and you can
list a number with each printed on a separate line.

The --rtems-bsp can be in new coming standard of arch/bsp such as
'sparc/sis' or you can just supply the bsp, eg 'sis'.