freebsd: Add support for FreeBSD 10.0.
GDB did not build out of the box on FreeBSD 10.0 due to the iconv move
to libc effecting the gettext port with Python. The patch provided is
from the gdb port (devel/gdb) and add the LDFLAGS to the link command line
when Python is used and this adds /usr/local/lib if needed.