Bugfix: gcc-4.4.7 couldn't be build with a modern texinfo package.
The texinfo documentation of binutils contained some constructs that are
errors, at least for a modern version of texinfo (version 6.0 and newer). This
meant that gcc-4.4.7 couldn't be build e.g. on fedora-22.

A patch file was added to source-builder/patches and an additional patch entry
was added to file rtems/config/tools/rtems-gcc-4.4.7-newlib-1.18.0-1.cfg. This
patch changes the problematic parts of the documentation.