Bugfix: The download url at ftp.rtems.org was wrong.
On the ftp site the directory structure has changed. This patch handles this
by a change in rtems/config/tools/rtems-kernel-4-1.cfg. Currently the 2-number
version number is created from a 3-number version number by a call to the sed
utility. Maybe there is is more elegant solution for this ?