Fix the rtems-tools waf --destdir path on Windows.
Waf requires the drive be added to the destdir path so it installs to a path
the RSB knows. The RSB assumes the destdir contains an MSYS drive path as
seen in the POSIX path. This is default for configure scripts.

Also fix the git URL to pull before the branch command as well as after to
update the branch.

Closes #2605.