rtems: Refactor to move to a single version on master.
Move the 4.12 version and base defines to just RTEMS base defines
which includes the version. This removes the version from a number of
files and makes it simpler to move to a new version. It however means
we cannot support more than version on master.

This change further breaks 4.10 and 4.9 however it means we can
branch these versions and better support them on their branches.