Fix GDB build on ArchLinux
Archlinux provides both, libguile v2.0 and v2.2. GDB states in
configuration its compatibility with both versions of libguile which is
false. The SCM_port interface of libguile was removed in v2.2 and
therefore breaks GDB as a user.

RTEMS does not use libguile and therefore it can be compiled without
support.

https://sourceware.org/bugzilla/show_bug.cgi?id=21104

Close #3054.