Avoid hard coded GCC binary paths
In order to build a tool chain with Ada support a native GCC with Ada
support of the same version as the cross compiler is required.  The RSB
used hard coded paths for the gcc and g++ programs.  This forced the RSB
user to change the main GCC installation of the machine to build a
particular RTEMS tool chain.  This is undesired/infeasible in most
situations.

Close #3284.