qemu: Update glib to build on newer gcc
Update the glib version to 2.48.2 in order to fix build failure on newer
gcc[0].

Force internal version of libpcre for glib, since otherwise the build
failed, presumably since libpcre is not provided as a module (maybe it
should?).

Increasing the glib version further to the current latest was not
tested.

[0]
  CC       libglib_2_0_la-gdate.lo
../../glib-2.39.3/glib/gdate.c: In function 'g_date_strftime':
../../glib-2.39.3/glib/gdate.c:2497:7: error: format not a string literal, format string not checked [-Werror=format-nonliteral]
       tmplen = strftime (tmpbuf, tmpbufsize, locale_format, &tm);
       ^~~~~~