sb/build: Add -E to '%source setup' to handle tar dependency errors.
Tar on Windows raises an error if a symlink appears before the target.
Windows needs the target to exist before a link can be created. This
patch adds support via the -E option to catch the error and rerun the
tar command a second time. The symlink should work as the target should
exist on the second pass.