rtems/config: Rename 8 to edge
What was formerly called 8 is now referred to as "edge". For example,
8/rtems-all is now edge/rtems-all. Additionally, edge tools now build as
the same version as normal tools builds, so care must be taken to
install them into a different directory than the standard tools install.
This was changed due to confusion of edge tools having a N+1 version
number and being perceived as better than the current version tools.

The explicit version number in the 7 tools has also been removed. This
is now sourced exclusively from version.py as part of the options
defaults.

Updates #92