newlib: Update to get POSIX Issue 8 prototypes (7 & next)
Change newlib hash from 5849108 to 038afec1.

This primarily is needed for including the recent additions for
POSIX Issue 8. It also picks up Sebastian Huber's addition of CLOCK_TAI,
a memcpy() fix for RISC-V to ensure the right implementation is used
for GCC 13, and fixes to MIPS newlib changes which resulted in
RTEMS not compiling for MIPS.

While updating this, an intervening commit to newlib removed ilp32.
The hash used reverts enought of that patch to include ilp32 again.

Updates gsoc#69
Updates #115