source-builder/sb: Prevent git downloads in no-download mode
Even with --no-download, git repositories were still being cloned. This
was because the git downloader didn't check for no-download mode and
simhost options didn't implement the download_disabled() check. This
fixes both issues and allows the no-download mode to operate properly
with git repositories. This also addresses a logic error in the file
downloader where an operand was inverted.

Closes #143